+++ examples/testrturtle.c
@@ -79,7 +79,8 @@
 
 int main() {
 	initWorld(800,600);
-	setSpeed(500);
+	setBackgroundImageFile("r:\\banner-001.png");
+	setSpeed(5000);
 	setPenSize(1);
 	
 	beginFill();